requirements.txt
gradio>=4.0
huggingface_hub>=0.20
langchain>=0.2
langchain_community
sentence-transformers>=3.0
faiss-cpu>=1.7
python-docx>=0.8
diskcache>=5.6
numpy>=1.26
matplotlib>=3.8
unstructured[md,pdf,docx]